+++ x/mint/module/abci.go
@@ -70,28 +70,6 @@ func BeginBlocker(ctx context.Context, k keeper.Keeper) error {
 	if err != nil {
 		return errorsmod.Wrap(err, "could not convert validators vs allora percent reward to legacy dec")
 	}
-	// every month on the first block of the month, update the emissions rate
-	// Note: Monthly rewards reset already happened above (before emission enabled check)
-	if blockCountSinceTGE%blocksPerMonth == 1 {
-		// Recalculate the target emission for the block
-		// WARNING: After Calling RecalculateTargetEmission,
-		// PreviousRewardEmissionPerUnitStakedToken and PreviousBlockEmission
-		// are set to new values. If later in begin blocker you need to use these
-		// you should get them first before this function is called!!
-		blockEmission, _, err = keeper.RecalculateTargetEmission(
-			sdkCtx,
-			k,
-			blockCountSinceTGE,
-			blocksPerMonth,
-			moduleParams,
-			ecosystemBalance,
-			ecosystemMintSupplyRemaining,
-			vPercent,
-		)
-		if err != nil {
-			return errorsmod.Wrap(err, "could not recalculate target emission")
-		}
-	}
 	// if the expected amount of emissions is greater than the balance of the ecosystem module account
 	if blockEmission.GT(ecosystemBalance) {
 		// mint the amount of tokens required to pay out the emissions