+++ x/mint/module/module_test.go
@@ -175,6 +175,12 @@ func (s *MintModuleTestSuite) SetupTest() {
 	s.appModule = mintAppModule
 }
 
+func (s *MintModuleTestSuite) recalculateEmission() cosmosMath.Int {
+	blockEmission, err := s.mintKeeper.RecalculateEmission(s.ctx)
+	s.Require().NoError(err)
+	return blockEmission
+}
+
 func TestMintModuleTestSuite(t *testing.T) {
 	suite.Run(t, new(MintModuleTestSuite))
 }
@@ -338,6 +344,8 @@ func (s *MintModuleTestSuite) TestNoNewMintedTokensIfInferenceRequestFeesEnoughT
 
 	tokenSupplyBefore := s.bankKeeper.GetSupply(s.ctx, sdk.DefaultBondDenom)
 
+	s.recalculateEmission()
+
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)
 
@@ -414,6 +422,8 @@ func (s *MintModuleTestSuite) TestTokensAreMintedIfInferenceRequestFeesNotEnough
 	s.Require().NoError(err)
 
 	tokenSupplyBefore := s.bankKeeper.GetSupply(s.ctx, sdk.DefaultBondDenom)
+
+	s.recalculateEmission()
 
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)
@@ -511,6 +521,8 @@ func (s *MintModuleTestSuite) TestNotEnoughTokensToMintToCoverInflation() {
 	s.Require().NoError(err)
 
 	tokenSupplyBefore := s.bankKeeper.GetSupply(s.ctx, sdk.DefaultBondDenom)
+
+	s.recalculateEmission()
 
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)
@@ -618,6 +630,8 @@ func (s *MintModuleTestSuite) TestInflationRateAsMorePeopleStakeGoesUp() {
 	ecosystemTokensMintedZero, err := s.mintKeeper.EcosystemTokensMinted.Get(s.ctx)
 	s.Require().NoError(err)
 	// do the first inflation calculation
+	s.recalculateEmission()
+
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)
 
@@ -649,6 +663,8 @@ func (s *MintModuleTestSuite) TestInflationRateAsMorePeopleStakeGoesUp() {
 	blocks := new(big.Int).SetUint64(blocksPerMonth)
 	blocks.Add(blocks, big.NewInt(1))
 	s.ctx = s.ctx.WithBlockHeight(blocks.Int64())
+
+	s.recalculateEmission()
 
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)
@@ -709,6 +725,8 @@ func (s *MintModuleTestSuite) TestEcosystemRefundReducesMintingInSubsequentBlock
 	tokenSupplyStart := s.bankKeeper.GetSupply(s.ctx, sdk.DefaultBondDenom)
 
 	// 2. Run BeginBlocker at block 1 - should mint tokens
+	s.recalculateEmission()
+
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)
 
@@ -742,6 +760,7 @@ func (s *MintModuleTestSuite) TestEcosystemRefundReducesMintingInSubsequentBlock
 
 	// 4. Run BeginBlocker at block 2
 	s.ctx = s.ctx.WithBlockHeight(2) // Advance block height, avoid recalculation
+
 	err = mint.BeginBlocker(s.ctx, s.mintKeeper)
 	s.Require().NoError(err)