+++ CHANGELOG.md
@@ -67,6 +67,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* [#881](https://github.com/allora-network/allora-chain/pull/881) Integrate `x/mint` with `x/scheduler` for periodic emission recalculation
+
 ### Changed
 
 ### Deprecated